Remove kai.eval.retrieval CLI; relocate shared probe primitives
Closes #672.

Delete src/kai/eval/retrieval.py and tests/test_eval_retrieval.py. The deprecated unscoped eval CLI no longer matches a production code path: every turn flows through assemble_turn_context when MEMORY_SCOPED_RECALL_ENABLED is on, so the CLI's metrics describe a path production no longer runs.

Relocate the shared probe primitives Probe, load_probes, probe_set_hash, detect_drift, and compute_rank into a new kai.eval._probes module so the behavioral evaluator and the collision-probe generator (which both depended on them) can keep working after retrieval.py is gone. Tests for the primitives move into tests/test_eval_probes.py.

Sweep docstrings and comments across kai.eval.retrieval_scoped, kai.eval.behavioral, kai.eval.gen_collision_probes, kai.eval._unscoped_recall_capture, and kai.eval.memory_backend_gate so they describe present state instead of comparing themselves to the deleted module. The retrieval_scoped harness prints 'production serves the unscoped pipeline' instead of 'LEGACY pipeline' when MEMORY_SCOPED_RECALL_ENABLED is off, since 'legacy' no longer points at a current module.

Verification: grep -r 'kai\.eval\.retrieval\b' src/ tests/ is empty; ruff clean; full pytest suite passes.
